Now correctly reproducing obfs4 test (description for details).
It is very strange how the implementation by Wang et al. worked, as the described "first client-to-server tcp payload" seems to be ANY non-empty TCP payload from the client to the server. It does seem to work with a high FPR when implemented as the reference implementation, but tweaking the p threshold and block size might help to reduce that." It still does not make much sense, but it works.